pyproject.toml
[tool.poetry]
name = "random-distributions"
version = "0.1.0"
description = "A Small Library of some Common Statistical Distributions"
authors = ["Nikhil Anand <mail@nikhil.io>"]

[tool.poetry.dependencies]
python = "^3.9"
seaborn = "^0.11.1"
pandas = "^1.2.4"
click = "^7.1.2"
progressbar2 = "^3.53.1"

[tool.poetry.scripts]
realize = "random_distributions.cli:entry_point"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
black = {version = "^21.5b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
